move LogsCentre into commons.logger
