+++ src/main/java/seedu/duke/task/command/TaskDeleteCommand.java
@@ -1,6 +1,6 @@
 package seedu.duke.task.command;
 
-import seedu.duke.CommandParseHelper;
+import seedu.duke.common.parser.CommandParseHelper;
 import seedu.duke.common.command.Command;
 import seedu.duke.common.model.Model;
 import seedu.duke.task.TaskList;
