+++ src/main/java/seedu/duke/task/entity/Task.java
@@ -1,6 +1,6 @@
 package seedu.duke.task.entity;
 
-import seedu.duke.CommandParseHelper;
+import seedu.duke.common.parser.CommandParseHelper;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
