+++ src/main/java/seedu/duke/ui/MainWindow.java
@@ -24,7 +24,7 @@ import javafx.scene.web.WebView;
 import javafx.stage.Popup;
 import javafx.stage.Screen;
 import javafx.stage.Stage;
-import seedu.duke.CommandParseHelper;
+import seedu.duke.common.parser.CommandParseHelper;
 
 import java.io.File;
 import java.io.FileInputStream;
